backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.3.0
reportlab==4.0.9
matplotlib==3.8.2
alembic==1.13.1
PyPDF2==3.0.1
pdfplumber==0.10.3
tabula-py==2.9.0
plaid-python==14.0.0
xmltodict==0.13.0
scikit-learn==1.3.2
scipy==1.11.4
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
